==> 07_Spectral_Analysis.Rmd <==
---
title: "07_Spectral_Analysis"
author: "Callum Weinberg"
date: "December 3, 2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Libraries

```{r}
library(dplyr, warn.conflicts = FALSE) #Using
library(tidyr) #Using
library(knitr) #Using
library(lubridate, warn.conflicts = FALSE) #Using
library(ggplot2) #Using
library(MASS) #Uncertain
library(qpcR) #Using
library(forecast) #Using
library(cowplot) #Using
library(TSA) #Using
```

##### Load Data #####

```{r}
load(file="Data/landings.Rdata")
landings_ts = landings$pounds

load(file="Data/landings_transformed_season_only.Rdata")
landing_ts_so = landings_transformed_season_only$pounds_transformed
```

###### Periodogram of Original Data #####
```{r}
## Periodogram
#periodogram = periodogram(landings$pounds, plot = TRUE)
periodogram = periodogram(landings$pounds, plot = FALSE)

# Put into Dataframe
periodogram_df = data.frame(periodogram = periodogram$spec[1:60], frequency = periodogram$freq[1:60])

# Plot
Periodogram_Graph = ggplot(data = periodogram_df, mapping = aes(x = frequency, y = periodogram)) +
  geom_segment(mapping = aes(xend = as.numeric(frequency), yend = 0)) +
  labs(x = "frequency", y = "periodogram") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 10),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 12),
    axis.text.x = element_text(size = 12),
    plot.title = element_text(hjust = 0.5, size = 12),
    axis.title=element_text(size=10,face="bold"))

Periodogram_Graph

# Get the Period at the spikes
print(1/periodogram_df[11,2])
print(1/periodogram_df[2,2])

# Fisher Test for Periodicity
library(GeneCycle) #Using
fisher.g.test(landings$pounds)

# Kolmogorov Smirnov Test
cpgram(landings$pounds,main="")



## Plots for report
png(filename = "Images/periodogram.png", width = 960, height = 480)
Periodogram_Graph
dev.off()
```

###### Periodogram of Residuals of Model 3 #####
```{r}
# Model 43 - Model Chosen for Forecasting Final Report. Corresponds to Model 3 of the report.
model43 = arima(landings_ts, order=c(1,0,1), seasonal = list(order = c(1,1,1), period = 12), 
                method = "ML",fixed = c(NA,NA,NA,NA))

## Periodogram
#periodogram = periodogram(landings$pounds, plot = TRUE)
periodogram = periodogram(residuals(model43))

# Put into Dataframe
periodogram_df = data.frame(periodogram = periodogram$spec[1:60], frequency = periodogram$freq[1:60])

# Plot
Periodogram_Residuals = ggplot(data = periodogram_df, mapping = aes(x = frequency, y = periodogram)) +
  geom_segment(mapping = aes(xend = as.numeric(frequency), yend = 0)) +
  labs(x = "frequency", y = "periodogram") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 10),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 12),
    axis.text.x = element_text(size = 12),
    plot.title = element_text(hjust = 0.5, size = 12),
    axis.title=element_text(size=10,face="bold"))

Periodogram_Residuals


## Plots for report
png(filename = "Images/periodogram_residuals.png", width = 960, height = 480)
Periodogram_Residuals
dev.off()
```


###### Tests for Periodicity #####

```{r}
# Fisher Test for Periodicity
library(GeneCycle) #Using
fisher.g.test(residuals(model43))

# Kolmogorov Smirnov Test
png(filename = "Images/KS_test Residuals.png", width = 480, height = 480)
cpgram(residuals(model43),main="")
dev.off()
````
